backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0
psycopg2-binary==2.9.9
groq==0.9.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
PyJWT==2.8.0
cryptography==42.0.8
aiohttp==3.9.5
twilio==9.0.5
faker==25.0.0
numpy==1.26.4
